kitchen_ros/src/LLM/requirements.txt
torch>=1.13
transformers>=4.30.0
accelerate>=0.21.0
opencv-python>=4.7.0
numpy>=1.23.0
pyzmq>=25.0.0
Pillow>=9.4.0
openai>=1.3.0
gradio>=4.0.0
typeguard>=2.13.0
catkin_pkg
rclpy